Correction article pour les floats et 1ere methode OK

+++ src/main/java/fr/exercice/service/ArticleService.java
@@ -33,7 +33,7 @@ public class ArticleService {
 		return this.ar.findByDescriptionContainingIgnoreCase(description);
 	}
 	
-	public List<Article> getByMarqueAndPrixRange(String marque, double prixMin, double prixMax){
+	public List<Article> getByMarqueAndPrixRange(String marque, Float prixMin, Float prixMax){
 		return this.ar.findByMarqueAndPrixRange(marque, prixMin, prixMax);
 	}
 }
